wasdeutsch/requirements.txt
Pillow>=10.0.0
edge-tts>=6.1.0
# Optional — faster, but not required:
# groq>=0.4.0   (uses urllib instead so no SDK needed)
